+++ src/lintrans/__main__.py
@@ -8,6 +8,7 @@
 
 """This module provides a :func:`main` function to interpret command line arguments and run the program."""
 
+from argparse import ArgumentParser
 import sys
 from textwrap import dedent
 
@@ -15,20 +16,35 @@ from lintrans import __version__
 from lintrans.gui import main_window
 
 
-def main(prog_name: str, args: list[str]) -> None:
+def main(args: list[str]) -> None:
     """Interpret program-specific command line arguments and run the main window in most cases.
 
     If the user supplies --help or --version, then we simply respond to that and then return.
     If they don't supply either of these, then we run :func:`lintrans.gui.main_window.main`.
 
-    ``prog_name`` is ``sys.argv[0]`` when this script is run with ``python -m lintrans``.
-
-    :param str prog_name: The name of the program
-    :param list[str] args: The other arguments to the program
+    :param list[str] args: The full argument list (including program name)
     """
-    if '-h' in args or '--help' in args:
-        print(dedent(f'''
-        Usage: {prog_name} [option]
+    parser = ArgumentParser(add_help=False)
+
+    parser.add_argument(
+        '-h',
+        '--help',
+        default=False,
+        action='store_true'
+    )
+
+    parser.add_argument(
+        '-V',
+        '--version',
+        default=False,
+        action='store_true'
+    )
+
+    parsed_args, unparsed_args = parser.parse_known_args()
+
+    if parsed_args.help:
+        print(dedent('''
+        Usage: lintrans [option]
 
         Options:
             -h, --help       Display this help text and exit
@@ -36,8 +52,9 @@ def main(prog_name: str, args: list[str]) -> None:
 
         Any other options will get passed to the QApplication constructor.
         If you don't know what that means, then don't provide any arguments and just the run the program.'''[1:]))
+        return
 
-    elif '-V' in args or '--version' in args:
+    if parsed_args.version:
         print(dedent(f'''
         lintrans (version {__version__})
         The linear transformation visualizer
@@ -46,10 +63,13 @@ def main(prog_name: str, args: list[str]) -> None:
 
         This program is licensed under GNU GPLv3, available here:
         <https://www.gnu.org/licenses/gpl-3.0.html>'''[1:]))
+        return
 
-    else:
-        main_window.main(args)
+    for arg in unparsed_args:
+        print(f'Passing "{arg}" to QApplication. See --help for recognised args')
+
+    main_window.main(args[:1] + unparsed_args)
 
 
 if __name__ == '__main__':
-    main(sys.argv[0], sys.argv[1:])
+    main(sys.argv)